53 login header live updates

+++ src/main/java/vv/pms/project/Project.java
@@ -57,7 +57,7 @@ public class Project {
     }
 
     public void setId(long id) {
-        this.id = id;
+        this.id = (Long) id;
     }
 
     public String getTitle() {
